Changed content_length to an unsigned long rather than unsigned int, fixes
2008-03-13  Jeffrey Stedfast  <fejj@novell.com>

	* gmime/gmime-parser.c (parser_construct_message): Changed
	content_length to an unsigned long rather than unsigned int, fixes
	bug #521872. Thanks to Pawel Salek for this fix.


svn path=/trunk/; revision=1227

+++ gmime/gmime-parser.c
@@ -1592,7 +1592,7 @@ static GMimeMessage *
 parser_construct_message (GMimeParser *parser)
 {
 	struct _GMimeParserPrivate *priv = parser->priv;
-	unsigned int content_length = ULONG_MAX;
+	unsigned long content_length = ULONG_MAX;
 	GMimeContentType *content_type;
 	struct _header_raw *header;
 	GMimeMessage *message;
